+++ lib/glimmer/ui/video.rb
@@ -1,7 +1,7 @@
 require 'glimmer/ui/custom_widget'
 require 'glimmer/swt/color_proxy'
 
-#TODO display progress wheel while loading video
+#TODO consider the need to display a progress wheel while loading video as an option
 
 module Glimmer
   module UI